requirements.txt
Django==5.2.3
djangorestframework==3.15.2
drf-spectacular==0.27.2
joblib==1.4.2
waitress==3.0.0

scikit-learn==1.5.2
pandas==2.2.2

xgboost==2.1.1

pandas==2.2.2

numpy==1.26.4
